Port quick phrase to temp mode
